fix: update updateWindGust method to use unsigned long and add interrupt guard for anemometerCounter

+++ index/Sensors.h
@@ -17,7 +17,7 @@ public:
     void readDHT(float &hum, float &temp);
     void readBMP(float &press, float &temp);
     void beginBMP();
-    void updateWindGust(unsigned int now);
+    void updateWindGust(unsigned long now); // Fix #2: was unsigned int — truncates millis()
     void reset();
     const Metrics &getMeasurements(unsigned long timestamp);
